Remove stray via, and disable zone fill
Zone fill needs to be re-added (maybe as part of DRC) before exporting.
Press "B" in the KiCad GUI.

Layout file will be more "stable" for version control without the zone fills,
and its size shrinks by more than a factor of two this way.
Of course the diff for this commit is huge.